tp4 completed

+++ app/src/main/java/com/insa/mygamelist/data/IGDB.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import com.insa.mygamelist.R
+import kotlinx.serialization.Contextual
 
 object IGDB {
 
@@ -49,7 +50,7 @@ data class Games(val id: Long,
                 val first_release_date: Long,
                 val genres: List<Long>,
                 val name: String,
-                val platforms: List<Int>,
+                val platforms: List<Long>,
                 val summary: String,
                 val total_rating: Float)
 data class Genre(val id: Long, val name: String)
